requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
av==14.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.5.0
dnspython==2.7.0
email_validator==2.2.0
enum34==1.1.10
fastapi==0.115.11
fastapi-cli==0.0.7
faster-whisper==1.1.1
filelock==3.17.0
flatbuffers==25.2.10
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-genai==1.4.0
h11==0.14.0
halo==0.0.31
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.2
humanfriendly==10.0
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
llvmlite==0.44.0
log-symbols==0.0.14
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.6.0
mpmath==1.3.0
narwhals==1.29.0
networkx==3.4.2
numba==0.61.0
numpy==1.26.4
onnxruntime==1.20.1
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
openwakeword==0.6.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
pvporcupine==1.9.5
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
RealtimeSTT==0.3.95
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.23.1
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.1
setuptools==75.8.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
spinners==0.0.24
starlette==0.46.0
streamlit==1.43.0
sympy==1.13.1
tenacity==9.0.0
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0
torchaudio==2.6.0
tornado==6.4.2
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
webrtcvad-wheels==2.0.14
websocket-client==1.8.0
websockets==14.2
